pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "NBA Text2SQL Knowledge Base — semantic retrieval layer for SQL generation."
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "pyyaml>=6.0",
]
